pyproject.toml
[project]
name = "pdf-type-router"
version = "1.0.0"
description = "PDF page classifier and extraction pipeline (digital vs scanned vs mixed)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymupdf>=1.23.0",
    "pdfplumber>=0.10.0",
    "pdfminer.six>=20221105",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "orjson>=3.9.0",
    "rapidfuzz>=3.0.0",
    "pdf2image>=1.16.0",
]

[project.optional-dependencies]
tables = [
    "camelot-py[cv]>=0.11.0",
    "tabula-py>=2.5.0",
]
ocr = [
    "paddleocr>=2.7.0",
    "pytesseract>=0.3.10",
]
mineru = [
    "mineru[all]>=2.7.0",
]
gpu = [
    "torch>=2.0.0",
    "paddlepaddle-gpu>=2.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
